feat: modular collect_data package and CI-friendly test layout
Introduce the ``collect_data`` package (config, storage, pipeline, ``python -m`` entry), keep the root ``collect_data.py`` shim, and document Parquet-first unified output plus server loading. Register the package in setuptools so editable installs and tox see ``collect_data`` imports. Normalize test and notebook ``sys.path`` setup with ``find_root()`` instead of cwd-relative paths. Fix a mypy issue in ``scrape_team_league_stats`` by separating Series vs one-row DataFrame names. Add checkpoint tracker unit tests.
