+++ collect_data/pipeline.py
@@ -1,0 +1,2312 @@
+"""Football data collection pipeline (scrapers + unified build).
+
+Sources:
+  FBref      (via ScraperFC / botasaurus Chrome) — basic stats for 8 leagues
+  Understat  (direct POST API, no browser)       — xG/xA for Big 5 leagues
+  SofaScore  (via ScraperFC / botasaurus Chrome) — 80+ stats for 37 leagues/comps
+
+Usage:
+  python3 collect_data.py                        # all sources + SofaScore matches + ClubElo + unified table
+  python3 -m collect_data --sofascore-only       # SofaScore season stats only
+  python3 -m collect_data --rebuild-only          # rebuild unified parquet from raw files
+"""
+
+from __future__ import annotations
+
+import argparse
+import concurrent.futures
+import json
+import logging
+import random
+import re
+import sys
+import time
+from datetime import datetime, timezone
+from io import StringIO
+from pathlib import Path
+
+import pandas as pd
+import requests
+
+from collect_data.config import (
+    FBREF_COL_RENAME,
+    FBREF_LEAGUES,
+    MANUAL_TM_OVERRIDES,
+    SEASONS,
+    SOFASCORE_SEASONS,
+    SOFASCORE_TARGET_LEAGUES,
+    SS_COL_RENAME,
+    STAT_CATEGORIES,
+    UNDERSTAT_AJAX_BASE,
+    UNDERSTAT_AJAX_HEADERS,
+    UNDERSTAT_API,
+    UNDERSTAT_AUTHORITATIVE,
+    UNDERSTAT_HEADERS,
+    UNDERSTAT_LEAGUES,
+)
+from collect_data.helpers import _norm_name, _silence_bota_noise, _ss_retry
+from collect_data.storage import (
+    DATA_DIR,
+    FRESHNESS_PATH,
+    RAW_DIR,
+    repo_root,
+    save_raw,
+    sofa_match_checkpoint_flush,
+    raw_freshness_age_hours,
+)
+
+log = logging.getLogger(__name__)
+
+# Re-scrape SofaScore *season* aggregates for the current season if older than this (hours).
+SOFASCORE_SEASON_REFRESH_HOURS = 168.0  # 7 days — suitable for weekly cron
+
+
+
+
+# ══════════════════════════════════════════════════════════════════════════════
+#  Shared helpers
+# ══════════════════════════════════════════════════════════════════════════════
+
+
+
+def fbref_season_to_understat(season: str) -> str:
+    """'2025-2026' → '2025'"""
+    return season.split("-")[0]
+
+
+# ══════════════════════════════════════════════════════════════════════════════
+#  FBref helpers
+# ══════════════════════════════════════════════════════════════════════════════
+
+def flatten_columns(df: pd.DataFrame) -> pd.DataFrame:
+    if not isinstance(df.columns, pd.MultiIndex):
+        return df
+    new_cols, seen = [], {}
+    for col in df.columns:
+        if isinstance(col, tuple):
+            parts = [str(p) for p in col if p and "Unnamed:" not in str(p)]
+            name = parts[-1] if parts else str(col[-1])
+        else:
+            name = str(col)
+        clean = FBREF_COL_RENAME.get(name, name)
+        if clean in seen:
+            if isinstance(col, tuple):
+                gp = [str(p) for p in col[:-1] if p and "Unnamed:" not in str(p)]
+                sfx = gp[0].lower().replace(" ","_").replace("+","_").replace("-","_") if gp else str(seen[clean])
+                clean = f"{clean}_{sfx}"
+            seen[clean] = seen.get(clean, 0) + 1
+        else:
+            seen[clean] = 0
+        new_cols.append(clean)
+    df = df.copy()
+    df.columns = new_cols
+    return df
+
+
+def clean_player_df(df: pd.DataFrame, league: str, season: str, stat_category: str) -> pd.DataFrame:
+    if df is None or df.empty:
+        return pd.DataFrame()
+    df = flatten_columns(df)
+    for col in ["player", "Player"]:
+        if col in df.columns:
+            df = df[df[col].notna() & ~df[col].astype(str).str.match(r"^(Player|Rk)$")]
+            break
+    skip = {"player","nation","pos","team","age","player_id","team_id","season","league","stat_category"}
+    for col in df.columns:
+        if col not in skip:
+            df[col] = pd.to_numeric(df[col], errors="coerce")
+    df["league"]        = league
+    df["season"]        = season
+    df["stat_category"] = stat_category
+    return df.reset_index(drop=True)
+
+
+# ══════════════════════════════════════════════════════════════════════════════
+#  FBref collection
+# ══════════════════════════════════════════════════════════════════════════════
+
+def collect_fbref(leagues=None, seasons=None, stat_categories=None, wait_time=7):
+    """Scrape FBref via ScraperFC (opens real Chrome browser)."""
+    sys.path.insert(0, str(repo_root() / "src"))
+    from ScraperFC import FBref
+
+    leagues         = leagues         or FBREF_LEAGUES
+    seasons         = seasons         or SEASONS
+    stat_categories = stat_categories or STAT_CATEGORIES
+
+    fbref = FBref(wait_time=wait_time)
+    total = len(leagues) * len(seasons) * len(stat_categories)
+    done  = 0
+
+    for season in seasons:
+        for league in leagues:
+            try:
+                valid = fbref.get_valid_seasons(league)
+            except Exception as e:
+                log.warning(f"Could not get valid seasons for {league}: {e}")
+                continue
+            if season not in valid:
+                log.warning(f"Season {season!r} not in FBref for {league}")
+                continue
+
+            for cat in stat_categories:
+                done += 1
+                fname = (f"{league.replace(' ','_')}__{season.replace('-','_')}"
+                         f"__{cat.replace(' ','_')}")
+                path  = RAW_DIR / f"{fname}.parquet"
+                if path.exists():
+                    log.info(f"[{done}/{total}] ⏭️  {path.name}")
+                    continue
+
+                log.info(f"[{done}/{total}] 📡 FBref: {league} / {season} / {cat}")
+                t0 = time.time()
+                try:
+                    result    = fbref.scrape_stats(season, league, cat)
+                    player_df = result.get("player", pd.DataFrame())
+                    if player_df is not None and not player_df.empty:
+                        player_df = clean_player_df(player_df, league, season, cat)
+                        save_raw(player_df, fname)
+                        log.info(f"  ✅ {len(player_df)} players in {time.time()-t0:.1f}s")
+                    else:
+                        log.warning("  ⚠️  No player data")
+                except Exception as e:
+                    log.error(f"  ❌ {e}")
+
+                elapsed = time.time() - t0
+                pause   = max(0, wait_time - elapsed) + random.uniform(0, 2)
+                if pause > 0:
+                    time.sleep(pause)
+
+
+# ══════════════════════════════════════════════════════════════════════════════
+#  Understat collection
+# ══════════════════════════════════════════════════════════════════════════════
+
+def collect_understat(leagues=None, seasons=None):
+    """Fetch per-player xG/xA from Understat's POST API (no browser needed)."""
+    leagues = leagues or list(UNDERSTAT_LEAGUES.keys())
+    seasons = seasons or SEASONS
+    RAW_DIR.mkdir(parents=True, exist_ok=True)
+
+    for league in leagues:
+        code = UNDERSTAT_LEAGUES.get(league)
+        if not code:
+            log.info(f"  Understat: skipping {league} (not Big 5)")
+            continue
+        for season in seasons:
+            us_year  = fbref_season_to_understat(season)
+            fname    = f"understat__{league.replace(' ','_')}__{season.replace('-','_')}"
+            raw_path = RAW_DIR / f"{fname}.parquet"
+            if raw_path.exists():
+                log.info(f"⏭️  {raw_path.name}")
+                continue
+            log.info(f"📡 Understat: {league} {season} (year={us_year})")
+            try:
+                resp = requests.post(UNDERSTAT_API,
+                                     data={"league": code, "season": us_year},
+                                     headers=UNDERSTAT_HEADERS, timeout=30)
+                resp.raise_for_status()
+                payload = resp.json()
+                if not payload.get("success"):
+                    log.warning(f"  Understat error: {payload}")
+                    continue
+                df = pd.DataFrame(payload["players"]).rename(columns={
+                    "player_name": "player", "time": "us_minutes", "games": "us_games",
+                    "goals": "us_goals", "xG": "xg", "assists": "us_assists", "xA": "xag",
+                    "shots": "us_shots", "key_passes": "us_key_passes",
+                    "yellow_cards": "us_yellow_cards", "red_cards": "us_red_cards",
+                    "position": "us_pos", "team_title": "us_team",
+                    "npg": "us_npg", "npxG": "npxg",
+                    "xGChain": "xg_chain", "xGBuildup": "xg_buildup",
+                })
+                num = ["us_minutes","us_games","us_goals","xg","us_assists","xag",
+                       "us_shots","us_key_passes","us_yellow_cards","us_red_cards",
+                       "us_npg","npxg","xg_chain","xg_buildup"]
+                for col in num:
+                    if col in df.columns:
+                        df[col] = pd.to_numeric(df[col], errors="coerce")
+                df["league"]     = league
+                df["season"]     = season
+                df["_name_norm"] = df["player"].apply(_norm_name)
+                save_raw(df, fname)
+                log.info(f"  ✅ {len(df)} players")
+            except Exception as e:
+                log.error(f"  ❌ Understat {league} {season}: {e}")
+            time.sleep(2)
+
+
+# ══════════════════════════════════════════════════════════════════════════════
+#  SofaScore collection
+# ══════════════════════════════════════════════════════════════════════════════
+
+def collect_sofascore(leagues=None, seasons=None):
+    """
+    Collect all player stats from SofaScore via ScraperFC.
+    Runs headless Chrome (botasaurus) — no visible browser window.
+    Covers 37 leagues/competitions including UCL, Europa League, Eredivisie, etc.
+    """
+    sys.path.insert(0, str(repo_root() / "src"))
+    from ScraperFC import Sofascore
+    from ScraperFC.utils import get_module_comps
+
+    all_ss_leagues = list(get_module_comps("SOFASCORE").keys())
+    leagues = leagues or SOFASCORE_TARGET_LEAGUES
+    seasons = seasons or SEASONS
+
+    RAW_DIR.mkdir(parents=True, exist_ok=True)
+    ss = Sofascore()
+
+    total = len(leagues) * len(seasons)
+    done  = 0
+
+    for league in leagues:
+        if league not in all_ss_leagues:
+            log.info(f"  SofaScore: {league!r} not supported, skipping")
+            continue
+
+        try:
+            with _silence_bota_noise():
+                valid = _ss_retry(
+                    ss.get_valid_seasons,
+                    league,
+                    label=f"get_valid_seasons({league})",
+                )
+        except Exception as e:
+            log.warning(f"  SofaScore get_valid_seasons failed for {league}: {e}")
+            continue
+
+        for season in seasons:
+            done += 1
+            ss_year = SOFASCORE_SEASONS.get(season)
+            if not ss_year or ss_year not in valid:
+                log.info(f"[{done}/{total}] ⏭️  {league} {season} — not in SofaScore ({ss_year!r})")
+                continue
+
+            fname    = f"sofascore__{league.replace(' ','_')}__{season.replace('-','_')}"
+            raw_path = RAW_DIR / f"{fname}.parquet"
+            if raw_path.exists():
+                if season != SEASONS[0]:
+                    log.info(f"[{done}/{total}] ⏭️  {raw_path.name}")
+                    continue
+                age_h = raw_freshness_age_hours(fname)
+                if age_h is not None and age_h < SOFASCORE_SEASON_REFRESH_HOURS:
+                    log.info(
+                        f"[{done}/{total}] ⏭️  {raw_path.name} "
+                        f"(current season, refreshed {age_h:.1f}h ago)"
+                    )
+                    continue
+
+            log.info(f"[{done}/{total}] 📡 SofaScore: {league} {season} ({ss_year})")
+            t0 = time.time()
+            try:
+                with _silence_bota_noise():
+                    df = ss.scrape_player_league_stats(ss_year, league, accumulation="total")
+
+                if df is None or df.empty:
+                    log.warning(f"  ⚠️  Empty result for {league} {season}")
+                    continue
+
+                # Rename columns
+                df = df.rename(columns=SS_COL_RENAME)
+                # Drop any duplicate columns that arise from rename collisions
+                df = df.loc[:, ~df.columns.duplicated(keep="first")]
+
+                # Numeric conversion for all stat columns
+                skip = {"player", "team", "league", "season", "_name_norm",
+                        "sofascore_id", "sofascore_team_id"}
+                for col in df.columns:
+                    if col not in skip:
+                        df[col] = pd.to_numeric(df[col], errors="coerce")
+
+                df["league"]     = league
+                df["season"]     = season
+                df["_name_norm"] = df["player"].apply(_norm_name)
+
+                save_raw(df, fname)
+                log.info(f"  ✅ {len(df)} players in {time.time()-t0:.1f}s")
+
+            except Exception as e:
+                log.error(f"  ❌ SofaScore {league} {season}: {e}")
+
+            time.sleep(1)
+
+
+# ══════════════════════════════════════════════════════════════════════════════
+#  SofaScore retry helper — mimics collect_sofascore's implicit "re-run to retry"
+# ══════════════════════════════════════════════════════════════════════════════
+
+
+
+# ══════════════════════════════════════════════════════════════════════════════
+#  ClubElo — global ratings + upcoming fixtures (plain HTTP, no browser)
+# ══════════════════════════════════════════════════════════════════════════════
+
+CLUBELO_HEADERS = {
+    "User-Agent": (
+        "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) "
+        "AppleWebKit/537.36 (KHTML, like Gecko) Chrome/120.0.0.0 Safari/537.36"
+    ),
+}
+
+
+def _clubelo_scoreline_probs(row: pd.Series) -> tuple[float, float, float]:
+    """Sum R:a-b columns into home win / draw / away win probabilities."""
+    hw, dr, aw = 0.0, 0.0, 0.0
+    for col, val in row.items():
+        if not isinstance(col, str) or not col.startswith("R:"):
+            continue
+        mid = col[2:]
+        if "-" not in mid:
+            continue
+        try:
+            a_str, b_str = mid.split("-", 1)
+            ha, aww = int(a_str), int(b_str)
+            p = float(val)
+        except (ValueError, TypeError):
+            continue
+        if ha > aww:
+            hw += p
+        elif ha == aww:
+            dr += p
+        else:
+            aw += p
+    return hw, dr, aw
+
+
+def collect_clubelo(date: str | None = None) -> None:
+    """
+    Fetch ClubElo global snapshot for one calendar day + upcoming fixtures CSV.
+    Uses http://api.clubelo.com (HTTPS often times out from some networks).
+    """
+    sys.path.insert(0, str(repo_root() / "src"))
+    from ScraperFC import ClubElo
+
+    RAW_DIR.mkdir(parents=True, exist_ok=True)
+    day = date or datetime.now(timezone.utc).strftime("%Y-%m-%d")
+    if not re.match(r"^\d{4}-\d{2}-\d{2}$", day):
+        log.error(f"  ClubElo: invalid date {day!r}, expected YYYY-MM-DD")
+        return
+
+    slug_date = day.replace("-", "_")
+    global_name = f"clubelo__global__{slug_date}"
+    fix_name = f"clubelo__fixtures__{slug_date}"
+    global_path = RAW_DIR / f"{global_name}.parquet"
+    fix_path = RAW_DIR / f"{fix_name}.parquet"
+
+    ce = ClubElo()
+
+    if not global_path.exists():
+        try:
+            log.info(f"📡 ClubElo global snapshot: {day}")
+            gdf = ce.scrape_date(day)
+            gdf = gdf.rename(
+                columns={
+                    "Rank": "rank",
+                    "Club": "club",
+                    "Country": "country",
+                    "Level": "level",
+                    "Elo": "elo",
+                    "From": "valid_from",
+                    "To": "valid_to",
+                }
+            )
+            gdf["snapshot_date"] = day
+            save_raw(gdf, global_name)
+        except Exception as e:
+            log.error(f"  ❌ ClubElo global {day}: {e}")
+    else:
+        log.info(f"⏭️  {global_name}.parquet already exists")
+
+    if not fix_path.exists():
+        try:
+            log.info("📡 ClubElo fixtures (upcoming)")
+            url = "http://api.clubelo.com/Fixtures"
+            r = requests.get(url, headers=CLUBELO_HEADERS, timeout=30)
+            r.raise_for_status()
+            fdf = pd.read_csv(StringIO(r.text))
+            probs = fdf.apply(_clubelo_scoreline_probs, axis=1, result_type="expand")
+            fdf = fdf.rename(
+                columns={
+                    "Date": "date",
+                    "Country": "country",
+                    "Home": "home_team",
+                    "Away": "away_team",
+                }
+            )
+            fdf["home_win_prob"] = probs[0]
+            fdf["draw_prob"] = probs[1]
+            fdf["away_win_prob"] = probs[2]
+            fdf["snapshot_date"] = day
+            keep = [
+                c
+                for c in fdf.columns
+                if c
+                in {
+                    "date",
+                    "country",
+                    "home_team",
+                    "away_team",
+                    "home_win_prob",
+                    "draw_prob",
+                    "away_win_prob",
+                    "snapshot_date",
+                }
+            ]
+            save_raw(fdf[keep], fix_name)
+            log.info(f"  ✅ ClubElo fixtures: {len(fdf)} rows")
+        except Exception as e:
+            log.error(f"  ❌ ClubElo fixtures: {e}")
+    else:
+        log.info(f"⏭️  {fix_name}.parquet already exists")
+
+
+# ══════════════════════════════════════════════════════════════════════════════
+#  SofaScore — per-match shots, team stats, player stats, momentum
+# ══════════════════════════════════════════════════════════════════════════════
+
+# SofaScore team-match-stat keys we persist (match-level; not duplicated in season parquet).
+_SS_TEAM_STAT_KEYS = {
+    "ballPossession": ("ball_possession", "pct"),
+    "expectedGoals": ("expected_goals", "float"),
+    "shotsOnGoal": ("shots_on_target", "float"),
+    "shotsOffGoal": ("shots_off_target", "float"),
+    "totalShotsOnGoal": ("total_shots", "float"),
+    "totalShotsInsideBox": ("shots_inside_box", "float"),
+    "totalShotsOutsideBox": ("shots_outside_box", "float"),
+    "finalThirdEntries": ("final_third_entries", "float"),
+    "touchesInOppBox": ("touches_in_opp_box", "float"),
+    "offsides": ("offsides", "float"),
+    "cornerKicks": ("corner_kicks", "float"),
+    "bigChanceCreated": ("big_chances", "float"),
+}
+
+
+def _sofascore_team_stats_to_rows(
+    team_df: pd.DataFrame,
+    match_id: int,
+    home_team: str,
+    away_team: str,
+    league: str,
+    season: str,
+) -> list[dict]:
+    """Pivot long SofaScore team match stats → one dict per period (ALL / 1ST / 2ND)."""
+    if team_df.empty or "period" not in team_df.columns or "key" not in team_df.columns:
+        return []
+    rows: list[dict] = []
+    for period in sorted(team_df["period"].unique()):
+        sub = team_df[team_df["period"] == period]
+        row: dict = {
+            "match_id": match_id,
+            "home_team": home_team,
+            "away_team": away_team,
+            "league": league,
+            "season": season,
+            "period": str(period),
+        }
+        for api_key, (prefix, _kind) in _SS_TEAM_STAT_KEYS.items():
+            hit = sub[sub["key"] == api_key]
+            if hit.empty:
+                row[f"{prefix}_home"] = None
+                row[f"{prefix}_away"] = None
+                continue
+            r0 = hit.iloc[0]
+            hv = r0.get("homeValue")
+            av = r0.get("awayValue")
+            # Prefer numeric homeValue/awayValue when present (SofaScore API).
+            row[f"{prefix}_home"] = float(hv) if pd.notna(hv) else None
+            row[f"{prefix}_away"] = float(av) if pd.notna(av) else None
+        rows.append(row)
+    return rows
+
+
+def _sofascore_shots_slim(shots: pd.DataFrame, match_id: int, league: str, season: str) -> pd.DataFrame:
+    """Normalize shotmap API frame to our slim schema."""
+    if shots.empty:
+        return pd.DataFrame()
+    out = shots.copy()
+    # Normalise xG column name from API
+    for cand in ("xg", "xG", "expectedGoals"):
+        if cand in out.columns and "xg" not in out.columns:
+            out = out.rename(columns={cand: "xg"})
+            break
+    for cand in ("xgot", "xGoT", "expectedGoalsOnTarget"):
+        if cand in out.columns and "xgot" not in out.columns:
+            out = out.rename(columns={cand: "xgot"})
+            break
+    if "player" in out.columns:
+        out["player_id"] = out["player"].apply(
+            lambda x: x.get("id") if isinstance(x, dict) else None
+        )
+        out["player_name"] = out["player"].apply(
+            lambda x: x.get("name") if isinstance(x, dict) else str(x)
+        )
+        out = out.drop(columns=["player"], errors="ignore")
+    coords = out.get("playerCoordinates")
+    if coords is not None:
+        out["player_x"] = coords.apply(lambda x: x.get("x") if isinstance(x, dict) else None)
+        out["player_y"] = coords.apply(lambda x: x.get("y") if isinstance(x, dict) else None)
+        out["player_z"] = coords.apply(lambda x: x.get("z") if isinstance(x, dict) else None)
+        out = out.drop(columns=["playerCoordinates"], errors="ignore")
+    gm = out.get("goalMouthCoordinates")
+    if gm is not None:
+        out["goal_mouth_x"] = gm.apply(lambda x: x.get("x") if isinstance(x, dict) else None)
+        out["goal_mouth_y"] = gm.apply(lambda x: x.get("y") if isinstance(x, dict) else None)
+        out["goal_mouth_z"] = gm.apply(lambda x: x.get("z") if isinstance(x, dict) else None)
+        out = out.drop(columns=["goalMouthCoordinates"], errors="ignore")
+    out["match_id"] = match_id
+    out["league"] = league
+    out["season"] = season
+    want = [
+        "match_id",
+        "player_id",
+        "player_name",
+        "isHome",
+        "time",
+        "shotType",
+        "situation",
+        "bodyPart",
+        "xg",
+        "xgot",
+        "goalMouthLocation",
+        "player_x",
+        "player_y",
+        "goal_mouth_x",
+        "goal_mouth_y",
+        "league",
+        "season",
+    ]
+    for c in want:
+        if c not in out.columns:
+            out[c] = None
+    slim = out[want].rename(
+        columns={
+            "isHome": "is_home",
+            "time": "minute",
+            "shotType": "shot_type",
+            "bodyPart": "body_part",
+            "goalMouthLocation": "goal_mouth_location",
+        }
+    )
+    return slim
+
+
+def _sofascore_player_match_slim(
+    pm: pd.DataFrame, match_id: int, home_id: int, league: str, season: str
+) -> pd.DataFrame:
+    """Keep only columns not already covered by season-level SofaScore parquet."""
+    if pm.empty:
+        return pd.DataFrame()
+    df = pm.loc[:, ~pm.columns.duplicated()].copy()
+    df["match_id"] = match_id
+    df["league"] = league
+    df["season"] = season
+    df["is_home"] = df["teamId"].apply(lambda tid: bool(tid == home_id) if pd.notna(tid) else None)
+    rename = {
+        "id": "player_id",
+        "name": "player_name",
+        "teamId": "team_id",
+        "teamName": "team_name",
+        "substitute": "substitute",
+        "minutesPlayed": "minutes_played",
+        "rating": "rating",
+        "expectedGoals": "xg",
+        "expectedGoalsOnTarget": "xgot",
+        "expectedAssists": "xa",
+        "goals": "goals",
+        "goalAssist": "assists",
+        "totalShots": "total_shots",
+        "onTargetScoringAttempt": "shots_on_target",
+        "totalPass": "total_passes",
+        "accuratePass": "accurate_passes",
+        "touches": "touches",
+        "possessionLostCtrl": "possession_lost",
+        "duelWon": "duels_won",
+        "duelLost": "duels_lost",
+        "fouls": "fouls",
+        "wasFouled": "was_fouled",
+    }
+    for old, new in rename.items():
+        if old in df.columns and new not in df.columns:
+            df[new] = df[old]
+    cols = [
+        "match_id",
+        "player_id",
+        "player_name",
+        "team_id",
+        "team_name",
+        "is_home",
+        "substitute",
+        "minutes_played",
+        "rating",
+        "xg",
+        "xgot",
+        "xa",
+        "goals",
+        "assists",
+        "total_shots",
+        "shots_on_target",
+        "total_passes",
+        "accurate_passes",
+        "touches",
+        "possession_lost",
+        "duels_won",
+        "duels_lost",
+        "fouls",
+        "was_fouled",
+        "league",
+        "season",
+    ]
+    for c in cols:
+        if c not in df.columns:
+            df[c] = None
+    return df[cols]
+
+
+def _ckpt_flush(
+    slug: str,
+    done_ids: set,
+    all_shots: list,
+    all_team_rows: list,
+    all_play: list,
+    all_mom: list,
+    p_shots: Path,
+    p_team: Path,
+    p_play: Path,
+    p_mom: Path,
+    p_ckpt: Path,
+    *,
+    last_match_date: str | None = None,
+    total_finished: int | None = None,
+) -> None:
+    """
+    Flush accumulated in-memory data to partial parquet files and write the
+    checkpoint JSON.  Called every FLUSH_EVERY matches so a crashed run can
+    resume from the last flush point rather than starting over.
+    """
+    sofa_match_checkpoint_flush(
+        slug,
+        done_ids,
+        all_shots,
+        all_team_rows,
+        all_play,
+        all_mom,
+        p_shots,
+        p_team,
+        p_play,
+        p_mom,
+        p_ckpt,
+        last_match_date=last_match_date,
+        total_finished=total_finished,
+    )
+
+
+def _sofascore_match_pack_fully_done(
+    force: bool,
+    p_shots: Path,
+    p_team: Path,
+    p_play: Path,
+    p_mom: Path,
+    p_ckpt: Path,
+    season: str | None = None,
+) -> bool:
+    """
+    Return True only when a league-season SofaScore match job can be skipped.
+
+    Checkpoint flushes write all four parquet files every FLUSH_EVERY matches, so
+    "all four exist" does **not** mean the season is finished — an in-progress run
+    always has ``sofascore_match_checkpoint__*.json`` alongside those files until
+    the worker removes it after the final save.
+    """
+    if force:
+        return False
+    if not (
+        p_shots.exists()
+        and p_team.exists()
+        and p_play.exists()
+        and p_mom.exists()
+    ):
+        return False
+    if p_ckpt.exists():
+        return False
+    # ``season is None`` = legacy callers / tests (treat as finished when files exist).
+    if season is not None and season == SEASONS[0]:
+        return False
+    return True
+
+
+def _discover_sofascore_match_jobs(
+    leagues: list[str] | None,
+    seasons: list[str] | None,
+    sleep_between_matches: float,
+    force: bool,
+) -> list[tuple[str, str, float, bool]]:
+    """
+    Build the list of (league, season, sleep_between_matches, force) tuples that
+    still need scraping.  Skips unsupported leagues, unavailable seasons, and
+    league-seasons that are fully finished (four parquets and no checkpoint),
+    unless force=True.
+    """
+    sys.path.insert(0, str(repo_root() / "src"))
+    from ScraperFC import Sofascore
+    from ScraperFC.utils import get_module_comps
+
+    all_ss_leagues = list(get_module_comps("SOFASCORE").keys())
+    use_leagues = leagues or list(SOFASCORE_TARGET_LEAGUES)
+    use_seasons = seasons or list(SEASONS)
+    RAW_DIR.mkdir(parents=True, exist_ok=True)
+    ss = Sofascore()
+    jobs: list[tuple[str, str, float, bool]] = []
+
+    for league in use_leagues:
+        if league not in all_ss_leagues:
+            log.info(f"  SofaScore matches: {league!r} not supported, skipping")
+            continue
+        try:
+            with _silence_bota_noise():
+                valid = _ss_retry(
+                    ss.get_valid_seasons,
+                    league,
+                    label=f"get_valid_seasons({league})",
+                )
+        except Exception as e:
+            log.warning(f"  SofaScore matches get_valid_seasons failed for {league}: {e}")
+            continue
+
+        for season in use_seasons:
+            ss_year = SOFASCORE_SEASONS.get(season)
+            if not ss_year or ss_year not in valid:
+                log.info(
+                    f"  SofaScore matches: {league} {season} — not available ({ss_year!r})"
+                )
+                continue
+
+            slug = f"{league.replace(' ', '_')}__{season.replace('-', '_')}"
+            p_shots = RAW_DIR / f"sofascore_match_shots__{slug}.parquet"
+            p_team = RAW_DIR / f"sofascore_match_team_stats__{slug}.parquet"
+            p_play = RAW_DIR / f"sofascore_match_player_stats__{slug}.parquet"
+            p_mom = RAW_DIR / f"sofascore_match_momentum__{slug}.parquet"
+            p_ckpt = RAW_DIR / f"sofascore_match_checkpoint__{slug}.json"
+            if _sofascore_match_pack_fully_done(
+                force, p_shots, p_team, p_play, p_mom, p_ckpt, season
+            ):
+                log.info(f"⏭️  SofaScore match pack complete for {slug}")
+                continue
+            jobs.append((league, season, sleep_between_matches, force))
+    return jobs
+
+
+def _sofascore_match_started_iso(m: dict) -> str | None:
+    """Best-effort kickoff time from a SofaScore ``events`` dict (UTC ISO)."""
+    ts = m.get("startTimestamp")
+    if not isinstance(ts, (int, float)) or ts <= 0:
+        return None
+    secs = float(ts) / 1000.0 if ts > 1_000_000_000_000 else float(ts)
+    return datetime.fromtimestamp(secs, tz=timezone.utc).isoformat()
+
+
+def _sofascore_ckpt_flush_meta(
+    done_ids: set[int], finished_by_id: dict[int, dict], total_finished: int
+) -> tuple[str | None, int]:
+    """Latest ISO start time among scraped IDs + API finished count."""
+    dates: list[str] = []
+    for mid in done_ids:
+        mm = finished_by_id.get(mid)
+        if not mm:
+            continue
+        iso = _sofascore_match_started_iso(mm)
+        if iso:
+            dates.append(iso)
+    last = max(dates) if dates else None
+    return last, int(total_finished)
+
+
+def _sofascore_bootstrap_done_ids_from_disk(
+    p_shots: Path, p_team: Path, p_play: Path, p_mom: Path
+) -> set[int]:
+    """Rebuild ``done_ids`` from existing match parquets (weekly incremental, no JSON)."""
+    ids: set[int] = set()
+    for p in (p_shots, p_team, p_play, p_mom):
+        if not p.exists():
+            continue
+        try:
+            df = pd.read_parquet(p, columns=["match_id"])
+            ids.update(df["match_id"].dropna().astype(int).tolist())
+        except Exception:
+            continue
+    return ids
+
+
+def _sofascore_match_season_worker(
+    job: tuple[str, str, float, bool],
+) -> None:
+    """
+    Scrape one league+season match pack (four parquets).  Top-level so it can be
+    pickled for ProcessPoolExecutor (macOS spawn).
+    """
+    league, season, sleep_between_matches, force = job
+    pfx = f"[{league} | {season}]"
+
+    sys.path.insert(0, str(repo_root() / "src"))
+    from ScraperFC import Sofascore
+    from ScraperFC.utils import get_module_comps
+
+    all_ss_leagues = list(get_module_comps("SOFASCORE").keys())
+    if league not in all_ss_leagues:
+        log.info(f"{pfx}  not supported, skipping")
+        return
+
+    ss = Sofascore()
+    try:
+        with _silence_bota_noise():
+            valid = _ss_retry(
+                ss.get_valid_seasons,
+                league,
+                label=f"get_valid_seasons({league})",
+            )
+    except Exception as e:
+        log.warning(f"{pfx}  get_valid_seasons failed: {e}")
+        return
+
+    ss_year = SOFASCORE_SEASONS.get(season)
+    if not ss_year or ss_year not in valid:
+        log.info(f"{pfx}  season not on SofaScore ({ss_year!r}), skipping")
+        return
+
+    slug = f"{league.replace(' ', '_')}__{season.replace('-', '_')}"
+    p_shots = RAW_DIR / f"sofascore_match_shots__{slug}.parquet"
+    p_team = RAW_DIR / f"sofascore_match_team_stats__{slug}.parquet"
+    p_play = RAW_DIR / f"sofascore_match_player_stats__{slug}.parquet"
+    p_mom = RAW_DIR / f"sofascore_match_momentum__{slug}.parquet"
+    p_ckpt = RAW_DIR / f"sofascore_match_checkpoint__{slug}.json"
+
+    if _sofascore_match_pack_fully_done(force, p_shots, p_team, p_play, p_mom, p_ckpt, season):
+        log.info(f"{pfx}  ⏭️  match pack already complete")
+        return
+
+    done_ids: set[int] = set()
+    if force:
+        _ckpt_flush(
+            slug,
+            set(),
+            [],
+            [],
+            [],
+            [],
+            p_shots,
+            p_team,
+            p_play,
+            p_mom,
+            p_ckpt,
+            last_match_date=None,
+            total_finished=0,
+        )
+        log.info(f"{pfx}  🔁 force-matches: checkpoint reset (was from any prior run)")
+    elif p_ckpt.exists():
+        try:
+            raw_ck = json.loads(p_ckpt.read_text(encoding="utf-8"))
+            done_ids = set(int(x) for x in raw_ck.get("done_ids", []))
+        except Exception:
+            pass
+    elif (
+        season == SEASONS[0]
+        and p_shots.exists()
+        and p_team.exists()
+        and p_play.exists()
+        and p_mom.exists()
+    ):
+        done_ids = _sofascore_bootstrap_done_ids_from_disk(p_shots, p_team, p_play, p_mom)
+        if done_ids:
+            log.info(
+                f"{pfx}  📂 incremental: {len(done_ids)} match ids loaded from existing parquets"
+            )
+
+    def _load_partial(path: Path) -> pd.DataFrame:
+        if path.exists() and done_ids:
+            try:
+                return pd.read_parquet(path)
+            except Exception:
+                pass
+        return pd.DataFrame()
+
+    all_shots: list[pd.DataFrame] = [df for df in [_load_partial(p_shots)] if not df.empty]
+    all_team_rows: list[dict] = (
+        _load_partial(p_team).to_dict("records") if p_team.exists() and done_ids else []
+    )
+    all_play: list[pd.DataFrame] = [df for df in [_load_partial(p_play)] if not df.empty]
+    all_mom: list[pd.DataFrame] = [df for df in [_load_partial(p_mom)] if not df.empty]
+
+    if done_ids:
+        log.info(f"{pfx}  📂 resuming: {len(done_ids)} matches already done")
+
+    log.info(f"{pfx}  📡 SofaScore matches ({ss_year})")
+    log.info(f"{pfx}  ⏳ fetching match list …")
+    try:
+        with _silence_bota_noise():
+            match_dicts = _ss_retry(
+                ss.get_match_dicts,
+                ss_year,
+                league,
+                label=f"get_match_dicts({league} {ss_year})",
+            )
+    except Exception as e:
+        log.error(f"{pfx}  ❌ get_match_dicts: {e}")
+        return
+
+    finished = [
+        m
+        for m in match_dicts
+        if isinstance(m.get("status"), dict) and m["status"].get("type") == "finished"
+    ]
+    finished_by_id: dict[int, dict] = {int(m["id"]): m for m in finished}
+    match_seq = {int(m["id"]): i for i, m in enumerate(finished, 1)}
+    remaining = [m for m in finished if int(m["id"]) not in done_ids]
+    log.info(
+        f"{pfx}  ✅ {len(finished)} finished matches total, "
+        f"{len(remaining)} remaining to scrape"
+    )
+
+    if not remaining:
+        log.info(f"{pfx}  all matches already done — writing final parquets")
+    else:
+        FLUSH_EVERY = 25
+
+        for idx, m in enumerate(remaining, 1):
+            mid = int(m["id"])
+            home = m.get("homeTeam", {}) or {}
+            away = m.get("awayTeam", {}) or {}
+            home_team = home.get("name", "")
+            away_team = away.get("name", "")
+            home_id = int(home.get("id", 0))
+
+            # global_idx = position in SofaScore's finished list for the season (not
+            # "how many matches scraped this session").  idx counts this run's queue.
+            global_idx = match_seq.get(mid, idx)
+            log.info(
+                f"{pfx}  [season {global_idx}/{len(finished)} · this run {idx}/{len(remaining)}] "
+                f"{home_team} vs {away_team} (id={mid})"
+            )
+
+            try:
+                with _silence_bota_noise():
+                    sdf = _ss_retry(ss.scrape_match_shots, mid, label=f"shots {mid}")
+                all_shots.append(_sofascore_shots_slim(sdf, mid, league, season))
+            except Exception as e:
+                log.warning(f"{pfx}    shots match {mid}: {e}")
+
+            try:
+                with _silence_bota_noise():
+                    tdf = _ss_retry(
+                        ss.scrape_team_match_stats, mid, label=f"team_stats {mid}"
+                    )
+                all_team_rows.extend(
+                    _sofascore_team_stats_to_rows(
+                        tdf, mid, home_team, away_team, league, season
+                    )
+                )
+            except Exception as e:
+                log.warning(f"{pfx}    team stats match {mid}: {e}")
+
+            try:
+                with _silence_bota_noise():
+                    pdf = _ss_retry(
+                        ss.scrape_player_match_stats, mid, label=f"player_stats {mid}"
+                    )
+                all_play.append(
+                    _sofascore_player_match_slim(pdf, mid, home_id, league, season)
+                )
+            except Exception as e:
+                log.warning(f"{pfx}    player stats match {mid}: {e}")
+
+            try:
+                with _silence_bota_noise():
+                    mdf = _ss_retry(ss.scrape_match_momentum, mid, label=f"momentum {mid}")
+                if not mdf.empty:
+                    mdf = mdf.copy()
+                    mdf["match_id"] = mid
+                    mdf["league"] = league
+                    mdf["season"] = season
+                    all_mom.append(mdf)
+            except Exception as e:
+                log.warning(f"{pfx}    momentum match {mid}: {e}")
+
+            done_ids.add(mid)
+            time.sleep(sleep_between_matches)
+
+            # Always flush once after the first match of this segment so Ctrl+C
+            # cannot leave a checkpoint from a *previous* run (see force-matches
+            # reset above) or an empty on-disk state for tens of matches.
+            if idx == 1 or idx % FLUSH_EVERY == 0:
+                log.info(f"{pfx}  💾 checkpoint flush at {global_idx}/{len(finished)} …")
+                lm, tf = _sofascore_ckpt_flush_meta(done_ids, finished_by_id, len(finished))
+                _ckpt_flush(
+                    slug,
+                    done_ids,
+                    all_shots,
+                    all_team_rows,
+                    all_play,
+                    all_mom,
+                    p_shots,
+                    p_team,
+                    p_play,
+                    p_mom,
+                    p_ckpt,
+                    last_match_date=lm,
+                    total_finished=tf,
+                )
+
+    shots_df = pd.concat(all_shots, ignore_index=True) if all_shots else pd.DataFrame()
+    team_df = pd.DataFrame(all_team_rows) if all_team_rows else pd.DataFrame()
+    play_df = pd.concat(all_play, ignore_index=True) if all_play else pd.DataFrame()
+    mom_df = pd.concat(all_mom, ignore_index=True) if all_mom else pd.DataFrame()
+
+    save_raw(shots_df, f"sofascore_match_shots__{slug}")
+    save_raw(team_df, f"sofascore_match_team_stats__{slug}")
+    save_raw(play_df, f"sofascore_match_player_stats__{slug}")
+    save_raw(mom_df, f"sofascore_match_momentum__{slug}")
+
+    if p_ckpt.exists():
+        p_ckpt.unlink()
+
+    log.info(
+        f"{pfx}  ✅ saved: shots={len(shots_df)} team_rows={len(team_df)} "
+        f"player_rows={len(play_df)} momentum_rows={len(mom_df)}"
+    )
+
+
+def collect_sofascore_matches(
+    leagues=None,
+    seasons=None,
+    sleep_between_matches: float = 2.0,
+    force: bool = False,
+    parallel: int = 1,
+) -> None:
+    """
+    Per-match SofaScore data: shot map, team stats (by half), player match stats, momentum.
+    Writes four parquet files per league+season. Skips when all four exist **and** there is
+    no checkpoint file (in-progress runs flush all four periodically). Use force=True to
+    rebuild from scratch.
+
+    parallel: if >1, run that many league-season jobs at once in separate processes
+    (safe with Bright Data — each job writes different files; IPs rotate per request).
+    """
+    RAW_DIR.mkdir(parents=True, exist_ok=True)
+    jobs = _discover_sofascore_match_jobs(leagues, seasons, sleep_between_matches, force)
+    if not jobs:
+        log.info("SofaScore matches: nothing to do (all packs complete or unavailable).")
+        return
+
+    log.info(f"SofaScore match jobs queued: {len(jobs)}  (parallel={parallel})")
+
+    if parallel <= 1:
+        for job in jobs:
+            _sofascore_match_season_worker(job)
+        return
+
+    with concurrent.futures.ProcessPoolExecutor(max_workers=parallel) as pool:
+        futures = {
+            pool.submit(_sofascore_match_season_worker, job): job for job in jobs
+        }
+        for fut in concurrent.futures.as_completed(futures):
+            job = futures[fut]
+            try:
+                fut.result()
+            except Exception as exc:
+                log.error(
+                    f"SofaScore match job crashed: {job[0]} {job[1]} — {exc}"
+                )
+
+
+# ══════════════════════════════════════════════════════════════════════════════
+#  Understat AJAX helpers
+# ══════════════════════════════════════════════════════════════════════════════
+
+def _us_ajax_get(path: str, referer: str | None = None) -> dict:
+    """GET request to a Understat AJAX endpoint with retries."""
+    headers = {**UNDERSTAT_AJAX_HEADERS, "Referer": referer or UNDERSTAT_AJAX_BASE}
+    for attempt in range(3):
+        try:
+            r = requests.get(UNDERSTAT_AJAX_BASE + path, headers=headers, timeout=20)
+            if r.status_code == 200 and len(r.content) > 10:
+                return r.json()
+        except Exception:
+            pass
+        time.sleep(attempt + 2)
+    return {}
+
+
+# ══════════════════════════════════════════════════════════════════════════════
+#  Understat — league tables
+# ══════════════════════════════════════════════════════════════════════════════
+
+def collect_understat_league_tables(leagues=None, seasons=None):
+    """Scrape overall/home/away league tables from Understat.  Big 5 only."""
+    leagues = leagues or list(UNDERSTAT_LEAGUES.keys())
+    seasons = seasons or SEASONS
+    RAW_DIR.mkdir(parents=True, exist_ok=True)
+
+    for league in leagues:
+        code = UNDERSTAT_LEAGUES.get(league)
+        if not code:
+            continue
+        for season in seasons:
+            us_year = fbref_season_to_understat(season)
+            slug    = f"{league.replace(' ','_')}__{season.replace('-','_')}"
+            overall_path = RAW_DIR / f"understat_league_table__{slug}__overall.parquet"
+            if overall_path.exists():
+                log.info(f"⏭️  understat_league_table__{slug}__overall.parquet")
+                continue
+
+            log.info(f"📡 Understat league table: {league} {season}")
+            data = _us_ajax_get(f"getLeagueData/{code}/{us_year}",
+                                referer=f"https://understat.com/league/{code}/{us_year}")
+            if not data:
+                log.warning(f"  ⚠️  No data for {league} {season}")
+                continue
+
+            teams = data.get("teams", {})
+            rows  = []
+            for t in teams.values():
+                for match in t.get("history", []):
+                    rows.append({
+                        "team_id":           t["id"],
+                        "team":              t["title"],
+                        "h_a":               match.get("h_a"),
+                        "xG":                float(match.get("xG", 0)),
+                        "xGA":               float(match.get("xGA", 0)),
+                        "npxG":              float(match.get("npxG", 0)),
+                        "npxGA":             float(match.get("npxGA", 0)),
+                        "ppda_att":          match.get("ppda", {}).get("att", 0),
+                        "ppda_def":          match.get("ppda", {}).get("def", 1),
+                        "ppda_allowed_att":  match.get("ppda_allowed", {}).get("att", 0),
+                        "ppda_allowed_def":  match.get("ppda_allowed", {}).get("def", 1),
+                        "deep":              int(match.get("deep", 0)),
+                        "deep_allowed":      int(match.get("deep_allowed", 0)),
+                        "goals":             int(match.get("scored", 0)),
+                        "goals_against":     int(match.get("missed", 0)),
+                        "xpts":              float(match.get("xpts", 0)),
+                        "wins":              int(match.get("wins", 0)),
+                        "draws":             int(match.get("draws", 0)),
+                        "losses":            int(match.get("loses", 0)),
+                        "pts":               int(match.get("pts", 0)),
+                        "npxGD":             float(match.get("npxGD", 0)),
+                        "date":              match.get("date"),
+                    })
+            if not rows:
+                continue
+
+            df = pd.DataFrame(rows)
+            agg = {
+                "wins": "sum", "draws": "sum", "losses": "sum", "pts": "sum",
+                "goals": "sum", "goals_against": "sum",
+                "xG": "sum", "xGA": "sum", "npxG": "sum", "npxGA": "sum", "npxGD": "sum",
+                "xpts": "sum", "deep": "sum", "deep_allowed": "sum",
+                "ppda_att": "sum", "ppda_def": "sum",
+                "ppda_allowed_att": "sum", "ppda_allowed_def": "sum",
+            }
+
+            def _build_table(sub):
+                t = sub.groupby(["team_id", "team"], as_index=False).agg(agg)
+                t["M"]     = t["wins"] + t["draws"] + t["losses"]
+                t["PPDA"]  = (t["ppda_att"] / t["ppda_def"].replace(0, 1)).round(2)
+                t["OPPDA"] = (t["ppda_allowed_att"] / t["ppda_allowed_def"].replace(0, 1)).round(2)
+                t = t.drop(columns=["ppda_att", "ppda_def", "ppda_allowed_att", "ppda_allowed_def"])
+                t["league"] = league
+                t["season"] = season
+                return t.sort_values("pts", ascending=False).reset_index(drop=True)
+
+            save_raw(_build_table(df),                      f"understat_league_table__{slug}__overall")
+            save_raw(_build_table(df[df["h_a"] == "h"]),   f"understat_league_table__{slug}__home")
+            save_raw(_build_table(df[df["h_a"] == "a"]),   f"understat_league_table__{slug}__away")
+            log.info(f"  ✅ {len(teams)} teams")
+            time.sleep(1)
+
+
+# ══════════════════════════════════════════════════════════════════════════════
+#  Understat — match shots + rosters
+# ══════════════════════════════════════════════════════════════════════════════
+
+def collect_understat_matches(leagues=None, seasons=None, sleep=0.5):
+    """Collect shot-level data and player-per-match rosters from Understat.  Big 5 only."""
+    leagues = leagues or list(UNDERSTAT_LEAGUES.keys())
+    seasons = seasons or SEASONS
+    RAW_DIR.mkdir(parents=True, exist_ok=True)
+
+    for league in leagues:
+        code = UNDERSTAT_LEAGUES.get(league)
+        if not code:
+            continue
+        for season in seasons:
+            us_year = fbref_season_to_understat(season)
+            slug    = f"{league.replace(' ','_')}__{season.replace('-','_')}"
+            shots_path = RAW_DIR / f"understat_match_shots__{slug}.parquet"
+            if shots_path.exists():
+                log.info(f"⏭️  understat_match_shots__{slug}.parquet")
+                continue
+
+            # Step 1: get match list for this season
+            log.info(f"📡 Understat matches: {league} {season} — fetching match IDs")
+            league_data = _us_ajax_get(f"getLeagueData/{code}/{us_year}",
+                                       referer=f"https://understat.com/league/{code}/{us_year}")
+            if not league_data:
+                log.warning(f"  ⚠️  No league data for {league} {season}")
+                continue
+
+            completed = [m for m in league_data.get("dates", []) if m.get("isResult")]
+            log.info(f"  {len(completed)} completed matches to fetch")
+
+            # Save match info from league data (fast, no extra requests)
+            info_path = RAW_DIR / f"understat_match_info__{slug}.parquet"
+            if not info_path.exists():
+                info_rows = []
+                for m in completed:
+                    info_rows.append({
+                        "match_id":      int(m["id"]),
+                        "home_team":     m["h"]["title"],
+                        "away_team":     m["a"]["title"],
+                        "home_team_id":  m["h"]["id"],
+                        "away_team_id":  m["a"]["id"],
+                        "home_goals":    int(m["goals"]["h"]),
+                        "away_goals":    int(m["goals"]["a"]),
+                        "home_xg":       float(m["xG"]["h"]),
+                        "away_xg":       float(m["xG"]["a"]),
+                        "datetime":      m.get("datetime"),
+                        "league":        league,
+                        "season":        season,
+                    })
+                save_raw(pd.DataFrame(info_rows), f"understat_match_info__{slug}")
+
+            # Step 2: fetch each match for shots + rosters
+            all_shots, all_rosters = [], []
+            for i, match in enumerate(completed, 1):
+                mid = match["id"]
+                if i % 50 == 0:
+                    log.info(f"  [{i}/{len(completed)}] fetching match {mid}")
+
+                mdata = _us_ajax_get(f"getMatchData/{mid}",
+                                     referer=f"https://understat.com/match/{mid}")
+                if not mdata:
+                    time.sleep(sleep)
+                    continue
+
+                # Shots
+                shots_raw = mdata.get("shots", {})
+                for h_a, shot_list in shots_raw.items():
+                    for s in (shot_list if isinstance(shot_list, list) else []):
+                        all_shots.append({
+                            "match_id":        int(mid),
+                            "player_id":       s.get("player_id"),
+                            "player":          s.get("player"),
+                            "h_a":             h_a,
+                            "minute":          s.get("minute"),
+                            "result":          s.get("result"),
+                            "X":               float(s.get("X", 0)),
+                            "Y":               float(s.get("Y", 0)),
+                            "xG":              float(s.get("xG", 0)),
+                            "situation":       s.get("situation"),
+                            "shot_type":       s.get("shotType"),
+                            "player_assisted": s.get("player_assisted"),
+                            "last_action":     s.get("lastAction"),
+                            "league":          league,
+                            "season":          season,
+                        })
+
+                # Rosters
+                for h_a, players in mdata.get("rosters", {}).items():
+                    for entry in (players.values() if isinstance(players, dict) else []):
+                        all_rosters.append({
+                            "match_id":        int(mid),
+                            "roster_entry_id": entry.get("id"),
+                            "player_id":       entry.get("player_id"),
+                            "player":          entry.get("player"),
+                            "team_id":         entry.get("team_id"),
+                            "position":        entry.get("position"),
+                            "h_a":             h_a,
+                            "minutes":         int(entry.get("time", 0)),
+                            "goals":           int(entry.get("goals", 0)),
+                            "own_goals":       int(entry.get("own_goals", 0)),
+                            "shots":           int(entry.get("shots", 0)),
+                            "xG":              float(entry.get("xG", 0)),
+                            "assists":         int(entry.get("assists", 0)),
+                            "xA":              float(entry.get("xA", 0)),
+                            "xGChain":         float(entry.get("xGChain", 0)),
+                            "xGBuildup":       float(entry.get("xGBuildup", 0)),
+                            "yellow_card":     int(entry.get("yellow_card", 0)),
+                            "red_card":        int(entry.get("red_card", 0)),
+                            "key_passes":      int(entry.get("key_passes", 0)),
+                            "roster_in":       entry.get("roster_in"),
+                            "roster_out":      entry.get("roster_out"),
+                            "position_order":  entry.get("positionOrder"),
+                            "league":          league,
+                            "season":          season,
+                        })
+
+                time.sleep(sleep)
+
+            if all_shots:
+                save_raw(pd.DataFrame(all_shots),   f"understat_match_shots__{slug}")
+            if all_rosters:
+                save_raw(pd.DataFrame(all_rosters), f"understat_rosters__{slug}")
+            log.info(f"  ✅ {len(all_shots)} shots, {len(all_rosters)} roster entries")
+
+
+# ══════════════════════════════════════════════════════════════════════════════
+#  Transfermarkt helpers + collection
+# ══════════════════════════════════════════════════════════════════════════════
+
+def _parse_tm_value(s: str) -> float | None:
+    """Parse '€45.00m' or '€500k' → float in EUR."""
+    if not isinstance(s, str):
+        return None
+    s = s.replace("€", "").replace(",", "").strip()
+    try:
+        if "m" in s.lower():
+            return float(s.lower().replace("m", "").strip()) * 1_000_000
+        if "k" in s.lower():
+            return float(s.lower().replace("k", "").strip()) * 1_000
+        return float(s)
+    except (ValueError, AttributeError):
+        return None
+
+
+def _tm_season(season: str, valid_seasons: dict) -> str | None:
+    """Map 'YYYY-YYYY' to whatever key Transfermarkt uses (e.g. '24/25')."""
+    start, end = season.split("-")
+    s2, e2 = start[2:], end[2:]
+    for fmt in [f"{s2}/{e2}", f"{start}/{e2}", f"{start}/{end}"]:
+        if fmt in valid_seasons:
+            return fmt
+    for k in valid_seasons:
+        if start in k or s2 in k:
+            return k
+    return None
+
+
+def collect_transfermarkt(leagues=None, seasons=None, sleep=2.0):
+    """Scrape player profiles from Transfermarkt for all supported leagues."""
+    sys.path.insert(0, str(repo_root() / "src"))
+    from ScraperFC import Transfermarkt
+    from ScraperFC.utils import get_module_comps
+
+    all_tm = list(get_module_comps("TRANSFERMARKT").keys())
+    leagues = leagues or all_tm
+    seasons = seasons or SEASONS
+    RAW_DIR.mkdir(parents=True, exist_ok=True)
+
+    tm = Transfermarkt()
+    for league in leagues:
+        if league not in all_tm:
+            log.info(f"  TM: {league!r} not supported, skipping")
+            continue
+
+        try:
+            valid = tm.get_valid_seasons(league)
+        except Exception as e:
+            log.warning(f"  TM get_valid_seasons failed for {league}: {e}")
+            continue
+
+        for season in seasons:
+            tm_year = _tm_season(season, valid)
+            if not tm_year:
+                log.info(f"  TM: {league} {season} not available (got {list(valid.keys())[:3]})")
+                continue
+
+            slug      = f"{league.replace(' ','_')}__{season.replace('-','_')}"
+            flat_path = RAW_DIR / f"transfermarkt__{slug}.parquet"
+            if flat_path.exists():
+                log.info(f"⏭️  transfermarkt__{slug}.parquet")
+                continue
+
+            log.info(f"📡 Transfermarkt: {league} {season} ({tm_year})")
+            try:
+                player_links = tm.get_player_links(tm_year, league)
+            except Exception as e:
+                log.error(f"  ❌ get_player_links: {e}")
+                continue
+
+            log.info(f"  {len(player_links)} players to scrape")
+            flat_rows, mv_rows, transfer_rows = [], [], []
+
+            for j, link in enumerate(player_links):
+                try:
+                    pdf = tm.scrape_player(link)
+                    if pdf.empty:
+                        continue
+                    row   = pdf.iloc[0]
+                    tm_id = str(row.get("ID", ""))
+
+                    flat_rows.append({
+                        "tm_id":               tm_id,
+                        "tm_name":             row.get("Name"),
+                        "market_value_str":    row.get("Value"),
+                        "market_value_eur":    _parse_tm_value(str(row.get("Value", ""))),
+                        "contract_expiration": row.get("Contract expiration"),
+                        "dob":                 row.get("DOB"),
+                        "age":                 row.get("Age"),
+                        "height_m":            row.get("Height (m)"),
+                        "nationality":         row.get("Nationality"),
+                        "citizenship":         json.dumps(row.get("Citizenship") or []),
+                        "tm_position":         row.get("Position"),
+                        "team":                row.get("Team"),
+                        "last_club":           row.get("Last club"),
+                        "joined":              row.get("Joined"),
+                        "since":               row.get("Since"),
+                        "league":              league,
+                        "season":              season,
+                        "_name_norm":          _norm_name(str(row.get("Name", ""))),
+                    })
+
+                    mvh = row.get("Market value history")
+                    if isinstance(mvh, pd.DataFrame) and not mvh.empty:
+                        for _, mr in mvh.iterrows():
+                            mv_rows.append({
+                                "tm_id": tm_id, "date": mr.get("date"),
+                                "value_eur": mr.get("value"),
+                                "league": league, "season": season,
+                            })
+
+                    th = row.get("Transfer history")
+                    if isinstance(th, pd.DataFrame) and not th.empty:
+                        for _, tr in th.iterrows():
+                            transfer_rows.append({
+                                "tm_id": tm_id, "tm_name": row.get("Name"),
+                                **{k: tr.get(k) for k in
+                                   ["Season", "Date", "Left", "Joined", "MV", "Fee"]},
+                                "league": league, "season": season,
+                            })
+
+                    if (j + 1) % 100 == 0:
+                        log.info(f"  [{j+1}/{len(player_links)}] scraped")
+
+                except Exception as e:
+                    log.warning(f"  ⚠️  {link}: {e}")
+
+                time.sleep(sleep)
+
+            if flat_rows:
+                save_raw(pd.DataFrame(flat_rows), f"transfermarkt__{slug}")
+                log.info(f"  ✅ {len(flat_rows)} players")
+            if mv_rows:
+                save_raw(pd.DataFrame(mv_rows), f"transfermarkt_mv_history__{slug}")
+            if transfer_rows:
+                save_raw(pd.DataFrame(transfer_rows), f"transfermarkt_transfers__{slug}")
+
+
+# ══════════════════════════════════════════════════════════════════════════════
+#  Capology helpers + collection
+# ══════════════════════════════════════════════════════════════════════════════
+
+def _cap_season(season: str, valid_seasons: list) -> str | None:
+    """Map 'YYYY-YYYY' to whatever string Capology uses (e.g. '2024-25')."""
+    start, end = season.split("-")
+    s2, e2 = start[2:], end[2:]
+    for fmt in [f"{s2}/{e2}", f"{start}/{e2}", f"{start}/{end}",
+                f"{start}-{e2}", f"{start}-{end}"]:
+        if fmt in valid_seasons:
+            return fmt
+    for vs in valid_seasons:
+        if start in vs or s2 in vs:
+            return vs
+    return None
+
+
+def _flatten_capology(df: pd.DataFrame) -> pd.DataFrame:
+    """Flatten Capology multi-index columns to snake_case strings."""
+    if not isinstance(df.columns, pd.MultiIndex):
+        return df
+    cols = []
+    for l0, l1 in df.columns:
+        l0c = l0.strip().lower().replace(" ", "_")
+        l1c = l1.strip().lower().replace(" ", "_")
+        if l1c and l1c != l0c:
+            cols.append(f"{l1c}__{l0c}".strip("_"))
+        else:
+            cols.append(l0c)
+    df = df.copy()
+    df.columns = cols
+    return df
+
+
+def collect_capology(leagues=None, seasons=None, currency="eur"):
+    """Scrape player wages from Capology for all supported leagues."""
+    sys.path.insert(0, str(repo_root() / "src"))
+    try:
+        from ScraperFC import Capology
+    except ImportError as e:
+        log.error(f"Capology import error (Selenium required): {e}")
+        return
+    from ScraperFC.utils import get_module_comps
+
+    all_cap = list(get_module_comps("CAPOLOGY").keys())
+    leagues  = leagues or all_cap
+    seasons  = seasons or SEASONS
+    RAW_DIR.mkdir(parents=True, exist_ok=True)
+
+    for league in leagues:
+        if league not in all_cap:
+            log.info(f"  Capology: {league!r} not supported, skipping")
+            continue
+
+        try:
+            valid = Capology().get_valid_seasons(league)
+        except Exception as e:
+            log.warning(f"  Capology get_valid_seasons failed for {league}: {e}")
+            continue
+
+        for season in seasons:
+            cap_year = _cap_season(season, valid)
+            if not cap_year:
+                log.info(f"  Capology: {league} {season} not available")
+                continue
+
+            slug = f"{league.replace(' ','_')}__{season.replace('-','_')}"
+            path = RAW_DIR / f"capology__{slug}.parquet"
+            if path.exists():
+                log.info(f"⏭️  capology__{slug}.parquet")
+                continue
+
+            log.info(f"📡 Capology: {league} {season} ({cap_year}) [{currency.upper()}]")
+            try:
+                df = Capology().scrape_salaries(cap_year, league, currency)
+                df = _flatten_capology(df)
+                df["league"]   = league
+                df["season"]   = season
+                df["currency"] = currency
+                save_raw(df, f"capology__{slug}")
+                log.info(f"  ✅ {len(df)} players")
+            except Exception as e:
+                log.error(f"  ❌ Capology {league} {season}: {e}")
+
+
+# ══════════════════════════════════════════════════════════════════════════════
+#  Financial merge (Transfermarkt + Capology → unified CSV)
+# ══════════════════════════════════════════════════════════════════════════════
+
+def _fuzzy_tm_fill(unmatched: pd.DataFrame, tm_lookup: pd.DataFrame,
+                   tm_cols: list[str], threshold: int = 85) -> pd.DataFrame:
+    """
+    For rows where the exact name merge missed, try rapidfuzz WRatio within
+    the same league+season bucket.  Only accepts matches >= threshold (0-100).
+    Short names (<= 5 chars) require >= 95 to avoid false positives like
+    'Kepa' matching 'Kepa Arrizabalaga' but not 'Igor Thiago' matching 'Thiago'.
+    Returns a DataFrame with the same index as unmatched, filled where possible.
+    """
+    from rapidfuzz import process, fuzz
+
+    filled = unmatched.copy()
+
+    for (league, season), group in unmatched.groupby(["league", "season"]):
+        bucket = tm_lookup[(tm_lookup["league"] == league) &
+                           (tm_lookup["season"] == season)]
+        if bucket.empty:
+            continue
+        tm_names  = bucket["_name_norm"].tolist()
+        tm_rows   = bucket.set_index("_name_norm")
+
+        for idx, row in group.iterrows():
+            query = row["_name_norm"]
+            cutoff = 95 if len(query) <= 5 else threshold
+            result = process.extractOne(
+                query, tm_names,
+                scorer=fuzz.WRatio,
+                score_cutoff=cutoff,
+            )
+            if result is None:
+                continue
+            best_name, score, _ = result
+            for col in tm_cols:
+                if col in tm_rows.columns:
+                    filled.at[idx, col] = tm_rows.at[best_name, col]
+
+    return filled
+
+
+def merge_financial_data(unified: pd.DataFrame) -> pd.DataFrame:
+    """Add market value, contract, wages to unified DataFrame from TM + Capology."""
+    unified = unified.copy()  # defragment before adding columns
+    unified["_name_norm"] = unified["player"].apply(_norm_name)
+
+    # ── Transfermarkt ─────────────────────────────────────────────────────────
+    # Pass 1: exact name match on name+league+season
+    # Pass 2: rapidfuzz WRatio fuzzy match for what's still unmatched
+    tm_files = [f for f in sorted(RAW_DIR.glob("transfermarkt__*.parquet"))
+                if "mv_history" not in f.name and "transfers" not in f.name]
+    if tm_files:
+        tm_frames = []
+        for f in tm_files:
+            try:
+                tm_frames.append(pd.read_parquet(f))
+            except Exception as e:
+                log.warning(f"Could not load {f.name}: {e}")
+        if tm_frames:
+            tm_df = pd.concat(tm_frames, ignore_index=True)
+            if "_name_norm" not in tm_df.columns:
+                tm_df["_name_norm"] = tm_df["tm_name"].apply(_norm_name)
+            tm_cols = [c for c in
+                       ["tm_id", "market_value_eur", "contract_expiration",
+                        "height_m", "nationality", "citizenship", "tm_position"]
+                       if c in tm_df.columns]
+            tm_lookup = (tm_df[["_name_norm", "league", "season"] + tm_cols]
+                         .sort_values("season", ascending=False)
+                         .drop_duplicates(subset=["_name_norm", "league", "season"], keep="first"))
+
+            # Pass 0 — manual overrides (remap unified name → TM name before merge)
+            for (nn, league_name), tm_nn in MANUAL_TM_OVERRIDES.items():
+                mask = (unified["_name_norm"] == nn) & (unified["league"] == league_name)
+                if mask.any():
+                    unified.loc[mask, "_name_norm"] = tm_nn
+
+            # Pass 1 — exact
+            unified = unified.merge(tm_lookup, on=["_name_norm", "league", "season"], how="left")
+            exact_matched = unified["tm_id"].notna().sum() if "tm_id" in unified.columns else 0
+
+            # Pass 2 — fuzzy for remaining unmatched rows
+            unmatched_mask = unified["tm_id"].isna()
+            if unmatched_mask.any():
+                filled = _fuzzy_tm_fill(
+                    unified[unmatched_mask][["_name_norm", "league", "season"] + tm_cols],
+                    tm_lookup, tm_cols,
+                )
+                for col in tm_cols:
+                    if col in filled.columns:
+                        unified.loc[unmatched_mask, col] = filled[col].values
+
+            total_matched = unified["tm_id"].notna().sum() if "tm_id" in unified.columns else 0
+            fuzzy_matched = total_matched - exact_matched
+            log.info(f"  TM merge: {total_matched}/{len(unified)} matched "
+                     f"({exact_matched} exact + {fuzzy_matched} fuzzy)")
+
+    # ── Capology ──────────────────────────────────────────────────────────────
+    cap_files = sorted(RAW_DIR.glob("capology__*.parquet"))
+    if cap_files:
+        cap_frames = []
+        for f in cap_files:
+            try:
+                cap_frames.append(pd.read_parquet(f))
+            except Exception as e:
+                log.warning(f"Could not load {f.name}: {e}")
+        if cap_frames:
+            cap_df = pd.concat(cap_frames, ignore_index=True)
+            # Find player name column (first non-league/season text col)
+            name_col = next(
+                (c for c in cap_df.columns
+                 if c not in {"league", "season", "currency"}
+                 and cap_df[c].dtype == object),
+                None,
+            )
+            if name_col:
+                cap_df["_name_norm"] = cap_df[name_col].apply(_norm_name)
+                wage_cols = [c for c in cap_df.columns if any(
+                    kw in c.lower() for kw in ["weekly", "annual", "gross", "wage", "salary"]
+                )]
+                log.info(f"  Capology wage columns: {wage_cols}")
+                if wage_cols:
+                    cap_merge = (
+                        cap_df[["_name_norm", "league", "season"] + wage_cols]
+                        .drop_duplicates(subset=["_name_norm", "league", "season"], keep="first")
+                    )
+                    unified = unified.merge(
+                        cap_merge, on=["_name_norm", "league", "season"], how="left"
+                    )
+                    log.info(f"  Capology merged wage columns: {wage_cols}")
+
+    unified = unified.drop(columns=["_name_norm"], errors="ignore")
+    return unified
+
+
+# ══════════════════════════════════════════════════════════════════════════════
+#  Load helpers
+# ══════════════════════════════════════════════════════════════════════════════
+
+_NON_FBREF_PREFIXES = (
+    "understat__", "understat_league_table__", "understat_match_info__",
+    "understat_match_shots__", "understat_rosters__",
+    "sofascore__", "sofascore_match_",
+    "clubelo__",
+    "transfermarkt__", "transfermarkt_mv_history__", "transfermarkt_transfers__",
+    "capology__",
+)
+
+
+def load_all_fbref_raw() -> pd.DataFrame:
+    frames = []
+    for f in sorted(RAW_DIR.glob("*.parquet")):
+        if f.name.startswith(_NON_FBREF_PREFIXES):
+            continue
+        try:
+            frames.append(pd.read_parquet(f))
+        except Exception as e:
+            log.warning(f"Could not load {f.name}: {e}")
+    return pd.concat(frames, ignore_index=True) if frames else pd.DataFrame()
+
+
+def load_all_understat_raw() -> pd.DataFrame:
+    frames = []
+    for f in sorted(RAW_DIR.glob("understat__*.parquet")):
+        try:
+            frames.append(pd.read_parquet(f))
+        except Exception as e:
+            log.warning(f"Could not load {f.name}: {e}")
+    return pd.concat(frames, ignore_index=True) if frames else pd.DataFrame()
+
+
+def load_all_sofascore_raw() -> pd.DataFrame:
+    frames = []
+    for f in sorted(RAW_DIR.glob("sofascore__*.parquet")):
+        try:
+            frames.append(pd.read_parquet(f))
+        except Exception as e:
+            log.warning(f"Could not load {f.name}: {e}")
+    return pd.concat(frames, ignore_index=True) if frames else pd.DataFrame()
+
+
+# ══════════════════════════════════════════════════════════════════════════════
+#  Build helpers
+# ══════════════════════════════════════════════════════════════════════════════
+
+def _build_base_from_understat(us_df: pd.DataFrame) -> pd.DataFrame:
+    """Build one row per (player, league, season) from Understat data."""
+    us_dedup = (
+        us_df.sort_values("us_minutes", ascending=False)
+        .drop_duplicates(subset=["_name_norm", "league", "season"], keep="first")
+        .copy()
+    )
+    rename = {
+        "id": "understat_id", "player": "player", "us_team": "team",
+        "us_pos": "pos", "us_games": "games", "us_minutes": "minutes",
+        "us_goals": "goals", "us_assists": "assists", "us_shots": "shots",
+        "us_key_passes": "key_passes", "us_yellow_cards": "yellow_cards",
+        "us_red_cards": "red_cards", "us_npg": "npg",
+        "league": "league", "season": "season",
+        "xg": "xg", "xag": "xag", "npxg": "npxg",
+        "xg_chain": "xg_chain", "xg_buildup": "xg_buildup",
+    }
+    cols = {k: v for k, v in rename.items() if k in us_dedup.columns}
+    base = us_dedup[list(cols)].rename(columns=cols).copy()
+    for col in ["games","minutes","goals","assists","shots","key_passes",
+                "yellow_cards","red_cards","npg","xg","xag","npxg",
+                "xg_chain","xg_buildup"]:
+        if col in base.columns:
+            base[col] = pd.to_numeric(base[col], errors="coerce").fillna(0)
+    base["ninety_s"] = (base["minutes"] / 90).round(2)
+    return base
+
+
+def _build_base_from_sofascore(ss_df: pd.DataFrame) -> pd.DataFrame:
+    """
+    Build one row per (player, league, season) from SofaScore data.
+    Used for non-Big5 leagues that have no Understat coverage.
+    """
+    if "_name_norm" not in ss_df.columns:
+        ss_df = ss_df.copy()
+        ss_df["_name_norm"] = ss_df["player"].apply(_norm_name)
+
+    ss_dedup = (
+        ss_df.sort_values("minutes" if "minutes" in ss_df.columns else ss_df.columns[0],
+                          ascending=False)
+        .drop_duplicates(subset=["_name_norm", "league", "season"], keep="first")
+        .copy()
+    )
+    return ss_dedup.drop(columns=["_name_norm", "sofascore_team_id"], errors="ignore")
+
+
+def _merge_understat_into(unified: pd.DataFrame, us_df: pd.DataFrame) -> pd.DataFrame:
+    """Merge Understat xG/xA columns into an existing FBref-based unified frame."""
+    if us_df.empty:
+        return unified
+    log.info(f"  Merging Understat: {len(us_df)} records, {us_df['league'].nunique()} leagues")
+    unified["_name_norm"] = unified["player"].apply(_norm_name)
+    if "_name_norm" not in us_df.columns:
+        us_df["_name_norm"] = us_df["player"].apply(_norm_name)
+
+    us_cols = ["xg","xag","npxg","xg_chain","xg_buildup","us_key_passes","us_npg","id"]
+    us_merge = (
+        us_df[["_name_norm","league","season"] + [c for c in us_cols if c in us_df.columns]]
+        .sort_values("xg", ascending=False)
+        .drop_duplicates(subset=["_name_norm","league","season"], keep="first")
+    )
+    unified = unified.merge(us_merge, on=["_name_norm","league","season"], how="left")
+
+    if "us_key_passes" in unified.columns:
+        if "key_passes" not in unified.columns:
+            unified = unified.rename(columns={"us_key_passes": "key_passes"})
+        else:
+            mask = unified["key_passes"].isna() | (unified["key_passes"] == 0)
+            unified.loc[mask, "key_passes"] = unified.loc[mask, "us_key_passes"]
+            unified = unified.drop(columns=["us_key_passes"])
+    if "us_npg" in unified.columns:
+        unified = unified.rename(columns={"us_npg": "npg"})
+    if "id" in unified.columns:
+        unified = unified.rename(columns={"id": "understat_id"})
+    unified = unified.drop(columns=["_name_norm"], errors="ignore")
+
+    matched = (unified["xg"].fillna(0) > 0).sum()
+    log.info(f"  xG coverage: {matched}/{len(unified)} ({100*matched//len(unified) if unified.shape[0] else 0}%)")
+    return unified
+
+
+def _merge_sofascore_into(unified: pd.DataFrame, ss_df: pd.DataFrame) -> pd.DataFrame:
+    """
+    Merge SofaScore stats into the unified DataFrame.
+
+    Strategy:
+    - For leagues already in unified (Big5 etc.): merge in NEW columns only.
+      Understat xG/goals/assists/minutes are kept; don't overwrite them.
+    - For leagues NOT in unified (Eredivisie, UCL, etc.): add new rows from SofaScore.
+    """
+    if ss_df.empty:
+        return unified
+
+    log.info(f"  Merging SofaScore: {len(ss_df)} records, {ss_df['league'].nunique()} leagues")
+
+    if "_name_norm" not in ss_df.columns:
+        ss_df = ss_df.copy()
+        ss_df["_name_norm"] = ss_df["player"].apply(_norm_name)
+
+    existing_leagues = set(unified["league"].unique())
+    ss_existing = ss_df[ss_df["league"].isin(existing_leagues)].copy()
+    ss_new      = ss_df[~ss_df["league"].isin(existing_leagues)].copy()
+
+    merge_key = ["_name_norm", "league", "season"]
+
+    # ── Part 1: merge new columns into existing rows ──────────────────────────
+    if not ss_existing.empty:
+        unified["_name_norm"] = unified["player"].apply(_norm_name)
+
+        # Which columns do we already have (keep them, don't overwrite)?
+        have_already = set(unified.columns) | UNDERSTAT_AUTHORITATIVE
+        ss_new_cols  = [c for c in ss_existing.columns
+                        if c not in have_already and c not in merge_key]
+
+        # Always bring the rating + defensive/aerial stats regardless
+        priority = [
+            "sofascore_id", "sofascore_team_id", "sofascore_rating",
+            "sofascore_rating_total", "sofascore_rating_count", "totw_appearances",
+            "starts", "tackles", "tackles_won", "tackles_won_pct",
+            "interceptions", "clearances", "blocked_shots", "outfield_blocks",
+            "errors_leading_to_goal", "errors_leading_to_shot", "dribbled_past",
+            "aerials_won", "aerials_won_pct", "aerials_lost",
+            "dribbles_completed", "dribbles_pct", "dribbles_attempted",
+            "ground_duels_won", "ground_duels_won_pct",
+            "duels_won", "duels_won_pct", "duels_lost",
+            "passes_total", "passes_completed", "passes_inaccurate", "pass_completion_pct",
+            "passes_final_third", "passes_opp_half", "passes_own_half",
+            "passes_opp_half_total", "passes_own_half_total",
+            "long_balls_total", "long_balls_completed", "long_balls_pct",
+            "crosses_total", "crosses_completed", "crosses_pct",
+            "chipped_passes_total", "chipped_passes_completed",
+            "pass_to_assist", "attempt_assists", "big_chances_created",
+            "big_chances_missed", "goals_inside_box", "goals_outside_box",
+            "goals_headed", "goals_left_foot", "goals_right_foot", "goals_penalty",
+            "goals_freekick", "own_goals", "goals_assists", "hit_woodwork",
+            "shots_inside_box", "shots_outside_box", "shots_on_target", "shots_off_target",
+            "blocked_shots", "shots_set_piece",
+            "goal_conversion_pct", "scoring_frequency", "set_piece_conversion",
+            "touches", "possession_lost", "possession_won_att_third",
+            "dispossessed", "ball_recoveries", "fouls", "fouled", "offsides",
+            "yellow_red_cards", "direct_red_cards",
+            "pens_taken", "pens_conceded", "pens_won", "pen_conversion_pct",
+            "pen_miss", "pen_post", "pen_on_target",
+            "saves", "saves_caught", "saves_parried",
+            "saves_inside_box", "saves_outside_box",
+            "goals_conceded", "goals_conceded_inside_box", "goals_conceded_outside_box",
+            "goals_prevented", "clean_sheets", "high_claims",
+            "crosses_not_claimed", "punches", "runs_out", "runs_out_successful",
+            "goal_kicks", "pens_saved", "pens_faced",
+        ]
+        bring = [c for c in priority if c in ss_existing.columns
+                 and c not in have_already]
+        # Also grab any other new columns we haven't listed
+        bring += [c for c in ss_new_cols if c not in bring]
+
+        if bring:
+            ss_merge = (
+                ss_existing[merge_key + bring]
+                .sort_values("sofascore_rating" if "sofascore_rating" in ss_existing.columns
+                             else bring[0], ascending=False)
+                .drop_duplicates(subset=merge_key, keep="first")
+            )
+            unified = unified.merge(ss_merge, on=merge_key, how="left")
+
+        unified = unified.drop(columns=["_name_norm"], errors="ignore")
+
+    # ── Part 2: add rows for new leagues (Eredivisie, UCL, etc.) ─────────────
+    if not ss_new.empty:
+        new_leagues = sorted(ss_new["league"].unique())
+        log.info(f"  Adding {len(new_leagues)} new leagues from SofaScore: {new_leagues}")
+        new_rows = _build_base_from_sofascore(ss_new)
+        # Ensure no duplicate columns before concat
+        new_rows = new_rows.loc[:, ~new_rows.columns.duplicated(keep="first")]
+        unified  = unified.loc[:, ~unified.columns.duplicated(keep="first")]
+        unified  = pd.concat([unified, new_rows], ignore_index=True)
+        log.info(f"  Unified now {len(unified)} rows after adding SofaScore leagues")
+
+    return unified
+
+
+def _finalize_and_save(
+    unified: pd.DataFrame, *, export_csv: bool = False
+) -> pd.DataFrame:
+    """Compute derived columns, fill NaN, save Parquet (+ optional CSV) + manifest."""
+    # Defragment before adding many new columns (avoids PerformanceWarning)
+    unified = unified.copy()
+
+    # Numeric fill
+    numeric_cols = unified.select_dtypes(include="number").columns
+    unified[numeric_cols] = unified[numeric_cols].fillna(0)
+
+    # ninety_s
+    if "ninety_s" not in unified.columns and "minutes" in unified.columns:
+        unified["ninety_s"] = (unified["minutes"] / 90).round(2)
+
+    n90 = unified.get("ninety_s", pd.Series(dtype=float)).replace(0, float("nan"))
+
+    per90_pairs = [
+        ("goals",               "goals_per90"),
+        ("assists",             "assists_per90"),
+        ("npg",                 "npg_per90"),
+        ("xg",                  "xg_per90"),
+        ("xag",                 "xag_per90"),
+        ("npxg",                "npxg_per90"),
+        ("xg_chain",            "xg_chain_per90"),
+        ("xg_buildup",          "xg_buildup_per90"),
+        ("shots",               "shots_per90"),
+        ("key_passes",          "key_passes_per90"),
+        ("tackles_won",         "tackles_won_per90"),
+        ("interceptions",       "interceptions_per90"),
+        ("big_chances_created", "big_chances_created_per90"),
+        ("dribbles_completed",  "dribbles_per90"),
+        ("progressive_passes",  "prog_passes_per90"),
+        ("progressive_carries", "prog_carries_per90"),
+        ("sca",                 "sca_per90"),
+        ("gca",                 "gca_per90"),
+    ]
+    for src, dst in per90_pairs:
+        if src in unified.columns and dst not in unified.columns:
+            unified[dst] = (pd.to_numeric(unified[src], errors="coerce") / n90).round(3).fillna(0)
+
+    # Aerial %
+    if "aerials_won" in unified.columns and "aerials_lost" in unified.columns:
+        if "aerials_won_pct" not in unified.columns or (unified["aerials_won_pct"] == 0).all():
+            tot = unified["aerials_won"] + unified["aerials_lost"]
+            unified["aerials_won_pct"] = (
+                unified["aerials_won"] / tot.replace(0, float("nan")) * 100
+            ).round(1).fillna(0)
+
+    # xG overperformance
+    if "goals" in unified.columns and "xg" in unified.columns:
+        unified["xg_overperformance"] = (
+            pd.to_numeric(unified["goals"], errors="coerce") -
+            pd.to_numeric(unified["xg"],   errors="coerce")
+        ).round(2).fillna(0)
+
+    if "npg" in unified.columns and "npxg" in unified.columns:
+        unified["npxg_overperformance"] = (
+            pd.to_numeric(unified["npg"],  errors="coerce") -
+            pd.to_numeric(unified["npxg"], errors="coerce")
+        ).round(2).fillna(0)
+
+    # Sort
+    sort_cols = [c for c in ["league","season","team","player"] if c in unified.columns]
+    if sort_cols:
+        unified = unified.sort_values(sort_cols).reset_index(drop=True)
+
+    # Save primary artifact as Parquet (columnar, typed, compact).
+    DATA_DIR.mkdir(parents=True, exist_ok=True)
+    out_pq = DATA_DIR / "unified_player_stats.parquet"
+    unified.to_parquet(out_pq, index=False)
+    log.info(
+        f"✅ Unified Parquet: {len(unified)} rows × {len(unified.columns)} cols → {out_pq}"
+    )
+    if export_csv:
+        out_csv = DATA_DIR / "unified_player_stats.csv"
+        unified.to_csv(out_csv, index=False)
+        log.info(f"  (also wrote CSV for spreadsheet use → {out_csv})")
+
+    manifest = {
+        "columns":       list(unified.columns),
+        "leagues":       sorted(unified["league"].unique().tolist()) if "league" in unified.columns else [],
+        "seasons":       sorted(unified["season"].unique().tolist()) if "season" in unified.columns else [],
+        "row_count":     len(unified),
+        "unified_path":  str(out_pq),
+        "xg_available":  bool("xg" in unified.columns and (unified["xg"] > 0).any()),
+        "rating_available": bool("sofascore_rating" in unified.columns and
+                                 (unified["sofascore_rating"] > 0).any()),
+        "last_built_at": datetime.now(timezone.utc).isoformat(),
+    }
+    oldest = None
+    if FRESHNESS_PATH.exists():
+        try:
+            fr = json.loads(FRESHNESS_PATH.read_text(encoding="utf-8"))
+            for _k, meta in fr.items():
+                ts = meta.get("fetched_at")
+                if not ts:
+                    continue
+                try:
+                    # fromisoformat accepts +00:00
+                    t = datetime.fromisoformat(ts.replace("Z", "+00:00"))
+                except ValueError:
+                    continue
+                oldest = t if oldest is None or t < oldest else oldest
+        except Exception as e:
+            log.warning(f"Could not read freshness for manifest: {e}")
+    manifest["oldest_source_fetched_at"] = oldest.isoformat() if oldest else None
+    (DATA_DIR / "manifest.json").write_text(json.dumps(manifest, indent=2))
+    log.info("📋 Manifest saved")
+    return unified
+
+
+# ══════════════════════════════════════════════════════════════════════════════
+#  Main build
+# ══════════════════════════════════════════════════════════════════════════════
+
+def build_unified(export_csv: bool = False):
+    """
+    Merge all raw data into data/unified_player_stats.parquet.
+
+    Layer order (each adds/overrides):
+      1. FBref   — basic stats, all leagues, if available
+      2. Understat — xG/xA for Big 5 leagues (overrides FBref xG when present)
+      3. SofaScore — 80+ stats for 37 leagues; adds rows for non-FBref leagues
+    """
+    log.info("🔨 Building unified player stats …")
+
+    fbref_df = load_all_fbref_raw()
+    us_df    = load_all_understat_raw()
+    ss_df    = load_all_sofascore_raw()
+
+    if fbref_df.empty and us_df.empty and ss_df.empty:
+        log.error("No raw data found. Run: python3 collect_data.py")
+        return pd.DataFrame()
+
+    key = ["player", "team", "league", "season"]
+
+    # ── Layer 1: FBref base ──────────────────────────────────────────────────
+    if not fbref_df.empty:
+        cats = list(fbref_df["stat_category"].unique())
+        log.info(f"  FBref categories: {sorted(cats)}")
+
+        base_cats  = ["standard", "playing time", "misc"]
+        other_cats = [c for c in cats if c not in base_cats]
+
+        def dedup(df):
+            if df.empty:
+                return df
+            df = df.copy()
+            df["_score"] = df.notna().sum(axis=1)
+            return (df.sort_values("_score", ascending=False)
+                      .drop_duplicates(subset=key, keep="first")
+                      .drop(columns=["_score"]))
+
+        base_frames = [dedup(fbref_df[fbref_df["stat_category"]==c]
+                             .drop(columns=["stat_category"], errors="ignore"))
+                       for c in base_cats if c in cats]
+        if not base_frames:
+            base_frames = [dedup(fbref_df[fbref_df["stat_category"]==cats[0]]
+                                 .drop(columns=["stat_category"], errors="ignore"))]
+
+        unified = base_frames[0]
+        for df in base_frames[1:]:
+            new_cols = [c for c in df.columns if c not in unified.columns or c in key]
+            unified  = unified.merge(df[new_cols], on=key, how="outer", suffixes=("","_dup"))
+            unified  = unified[[c for c in unified.columns if not c.endswith("_dup")]]
+
+        for cat in other_cats:
+            cat_df = dedup(fbref_df[fbref_df["stat_category"]==cat]
+                           .drop(columns=["stat_category","nation","pos","age","born",
+                                          "player_id","team_id"], errors="ignore"))
+            if cat_df.empty:
+                continue
+            new_cols = [c for c in cat_df.columns if c not in unified.columns or c in key]
+            if len(new_cols) <= len(key):
+                continue
+            unified = unified.merge(cat_df[new_cols], on=key, how="left", suffixes=("","_dup"))
+            unified = unified[[c for c in unified.columns if not c.endswith("_dup")]]
+            log.info(f"  FBref '{cat}': {len(unified)} rows, {len(unified.columns)} cols")
+
+    elif not us_df.empty:
+        # No FBref — build base from Understat
+        unified = _build_base_from_understat(us_df)
+        us_df   = pd.DataFrame()  # already consumed
+
+    elif not ss_df.empty:
+        # No FBref, no Understat — build entirely from SofaScore
+        log.info("  No FBref/Understat — building base from SofaScore only")
+        unified = _build_base_from_sofascore(ss_df)
+        ss_df   = pd.DataFrame()  # already consumed
+
+    # ── Layer 2: Understat xG ────────────────────────────────────────────────
+    if not us_df.empty:
+        unified = _merge_understat_into(unified, us_df)
+
+    # ── Layer 3: SofaScore ───────────────────────────────────────────────────
+    if not ss_df.empty:
+        unified = _merge_sofascore_into(unified, ss_df)
+
+    # ── Layer 4: Financial data (Transfermarkt + Capology) ───────────────────
+    unified = merge_financial_data(unified)
+
+    return _finalize_and_save(unified, export_csv=export_csv)
+
+
+# ══════════════════════════════════════════════════════════════════════════════
+#  CLI
+# ══════════════════════════════════════════════════════════════════════════════
+
+
+def main() -> None:
+    p = argparse.ArgumentParser(
+        description="Collect football stats — FBref + Understat + SofaScore + Transfermarkt + Capology"
+    )
+    p.add_argument("--leagues",                 nargs="*", help="Override league list")
+    p.add_argument("--seasons",                 nargs="*", help="Seasons e.g. 2025-2026 2024-2025")
+    p.add_argument("--stats",                   nargs="*", help="FBref stat categories")
+    p.add_argument("--wait",                    type=int,   default=7,   help="FBref request delay (s)")
+    p.add_argument("--sleep",                   type=float, default=0.5, help="Sleep between requests for TM / Understat matches")
+    p.add_argument("--export-csv",              action="store_true",
+                   help="Also write unified_player_stats.csv alongside the Parquet build")
+    # Source flags
+    p.add_argument("--fbref-only",              action="store_true")
+    p.add_argument("--understat-only",          action="store_true", help="Run old Understat season stats only")
+    p.add_argument("--understat-tables-only",   action="store_true", help="Run Understat league tables only (fast)")
+    p.add_argument("--understat-matches-only",  action="store_true", help="Run Understat match shots + rosters only (~90 min)")
+    p.add_argument("--sofascore-only",          action="store_true")
+    p.add_argument("--sofascore-matches-only",  action="store_true",
+                   help="SofaScore per-match shots / team / player / momentum only (~hours)")
+    p.add_argument("--clubelo-only",            action="store_true",
+                   help="ClubElo global ratings + fixtures (HTTP, seconds)")
+    p.add_argument("--clubelo-date",            type=str, default=None,
+                   help="YYYY-MM-DD for ClubElo global snapshot (default: today UTC)")
+    p.add_argument("--force-matches",             action="store_true",
+                   help="Re-fetch SofaScore match parquets even if all four exist")
+    p.add_argument(
+        "--parallel",
+        type=int,
+        default=1,
+        metavar="N",
+        help=(
+            "SofaScore per-match: run up to N league-season jobs in parallel processes "
+            "(default: 1). Safe with Bright Data — each job writes its own parquet files."
+        ),
+    )
+    p.add_argument("--transfermarkt-only",      action="store_true", help="Run Transfermarkt player profiles only (~6 hrs)")
+    p.add_argument("--capology-only",           action="store_true", help="Run Capology wages only (~2 hrs)")
+    p.add_argument("--rebuild-only",            action="store_true", help="Skip scraping; rebuild unified table from raw files")
+    # Skip flags
+    p.add_argument("--no-fbref",                action="store_true")
+    p.add_argument("--no-understat",            action="store_true")
+    p.add_argument("--no-sofascore",            action="store_true")
+    p.add_argument("--no-sofascore-matches",    action="store_true",
+                   help="Skip SofaScore per-match collection on full runs")
+    p.add_argument("--no-clubelo",              action="store_true", help="Skip ClubElo snapshot on full runs")
+    p.add_argument("--no-transfermarkt",        action="store_true")
+    p.add_argument("--no-capology",             action="store_true")
+    args = p.parse_args()
+
+    leagues = args.leagues or None
+    seasons = args.seasons or SEASONS
+    stats   = args.stats   or STAT_CATEGORIES
+
+    only_flags = [
+        ("fbref",             args.fbref_only),
+        ("understat",         args.understat_only),
+        ("understat_tables",  args.understat_tables_only),
+        ("understat_matches", args.understat_matches_only),
+        ("sofascore",         args.sofascore_only),
+        ("sofascore_matches", args.sofascore_matches_only),
+        ("clubelo",           args.clubelo_only),
+        ("transfermarkt",     args.transfermarkt_only),
+        ("capology",          args.capology_only),
+    ]
+    active_only = next((name for name, flag in only_flags if flag), None)
+
+    def _run(source: str) -> bool:
+        if active_only:
+            return source == active_only
+        skip_attr = f"no_{source.split('_')[0]}"
+        return not getattr(args, skip_attr, False)
+
+    print("=" * 70)
+    print("  Football Data Collector")
+    print("=" * 70)
+    print(f"Seasons : {seasons}")
+    if leagues:
+        print(f"Leagues : {leagues}")
+    print()
+
+    if not args.rebuild_only:
+        if _run("understat"):
+            print("── Understat (season xG/xA, Big 5) ─────────────────────────────────")
+            collect_understat(leagues=leagues, seasons=seasons)
+            print()
+
+        if _run("understat_tables"):
+            print("── Understat league tables (xG, PPDA, Big 5) ───────────────────────")
+            collect_understat_league_tables(leagues=leagues, seasons=seasons)
+            print()
+
+        if _run("understat_matches"):
+            print("── Understat match shots + rosters (Big 5) ─────────────────────────")
+            collect_understat_matches(leagues=leagues, seasons=seasons, sleep=args.sleep)
+            print()
+
+        if _run("sofascore"):
+            print("── SofaScore (80+ stats, target leagues) ───────────────────────────")
+            collect_sofascore(leagues=leagues, seasons=seasons)
+            print()
+
+        run_sofascore_matches = (
+            active_only == "sofascore_matches"
+            or (
+                active_only is None
+                and not args.no_sofascore_matches
+                and not args.no_sofascore
+            )
+        )
+        if run_sofascore_matches:
+            print("── SofaScore per-match (shots, team stats, player stats, momentum) ─")
+            pw = max(1, int(args.parallel))
+            if pw > 1:
+                print(f"Parallel workers : {pw}")
+            collect_sofascore_matches(
+                leagues=leagues,
+                seasons=seasons,
+                force=args.force_matches,
+                parallel=pw,
+            )
+            print()
+
+        run_clubelo = active_only == "clubelo" or (
+            active_only is None and not args.no_clubelo
+        )
+        if run_clubelo:
+            print("── ClubElo (ratings + fixtures) ─────────────────────────────────────")
+            collect_clubelo(date=args.clubelo_date)
+            print()
+
+        if _run("fbref"):
+            print("── FBref (basic stats, 8 leagues) ──────────────────────────────────")
+            collect_fbref(leagues=leagues or FBREF_LEAGUES,
+                          seasons=seasons, stat_categories=stats, wait_time=args.wait)
+            print()
+
+        if _run("transfermarkt"):
+            print("── Transfermarkt (player profiles + MV, all supported leagues) ─────")
+            collect_transfermarkt(leagues=leagues, seasons=seasons, sleep=args.sleep)
+            print()
+
+        if _run("capology"):
+            print("── Capology (wages EUR, all supported leagues) ──────────────────────")
+            collect_capology(leagues=leagues, seasons=seasons, currency="eur")
+            print()
+
+    supplementary_only = active_only in (
+        "understat_tables", "understat_matches", "transfermarkt", "capology",
+        "sofascore_matches", "clubelo",
+    )
+    if not supplementary_only:
+        print("── Building unified Parquet ─────────────────────────────────────────")
+        build_unified(export_csv=args.export_csv)
+    else:
+        print(f"── Skipping unified rebuild (supplementary data only: {active_only}) ──")
+        print("   Run: python3 -m collect_data --rebuild-only  to merge into unified table")
+    print("\nDone! ✅")
