+++ test/test_clubelo.py
@@ -2,8 +2,10 @@ import sys
 import pandas as pd
 import pytest
 from contextlib import nullcontext as does_not_raise
+from rootutils import find_root
 
-sys.path.append('./src/')
+# Resolve ``src`` from the repo root so tests work regardless of cwd (no hard-coded paths).
+sys.path.append(str(find_root() / "src"))
 from ScraperFC import ClubElo
